Added planning Driver / Python plan generator ongoing

+++ src/route2_table_t.hpp
@@ -67,13 +67,13 @@ class route2_table_t
         inline int _encodeNoHash(void *buf, int offset, int maxlen) const;
         inline int _getEncodedSizeNoHash() const;
         inline int _decodeNoHash(const void *buf, int offset, int maxlen);
-        inline static int64_t _computeHash(const __lcm_hash_ptr *p);
+        inline static uint64_t _computeHash(const __lcm_hash_ptr *p);
 };
 
 int route2_table_t::encode(void *buf, int offset, int maxlen) const
 {
     int pos = 0, tlen;
-    int64_t hash = getHash();
+    int64_t hash = (int64_t)getHash();
 
     tlen = __int64_t_encode_array(buf, offset + pos, maxlen - pos, &hash, 1);
     if(tlen < 0) return tlen; else pos += tlen;
@@ -168,7 +168,7 @@ int route2_table_t::_getEncodedSizeNoHash() const
     return enc_size;
 }
 
-int64_t route2_table_t::_computeHash(const __lcm_hash_ptr *p)
+uint64_t route2_table_t::_computeHash(const __lcm_hash_ptr *p)
 {
     const __lcm_hash_ptr *fp;
     for(fp = p; fp != NULL; fp = fp->parent)
@@ -176,7 +176,7 @@ int64_t route2_table_t::_computeHash(const __lcm_hash_ptr *p)
             return 0;
     const __lcm_hash_ptr cp = { p, (void*)route2_table_t::getHash };
 
-    int64_t hash = 0x7736c24c73b46136LL +
+    uint64_t hash = 0x7736c24c73b46136LL +
          route2_entry_t::_computeHash(&cp);
 
     return (hash<<1) + ((hash>>63)&1);
